pyproject.toml
[project]
name = "agentic-sdlc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=1.0.0",
    "aiosqlite>=0.22.1",
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-mcp-adapters>=0.2.2",
    "httpx>=0.27.0",
    "requests>=2.32.0",
    "tavily-python>=0.5.0",
    "langchain-openai>=1.1.14",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.1.8",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "loguru>=0.7.3",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "streamlit>=1.56.0",
    "uvicorn[standard]>=0.44.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-proto>=1.27.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
    "fpdf2>=2.7.9",
    "markdown>=3.6",
    "pytest>=8.3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
